% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyNDB.R
\name{save_parquet}
\alias{save_parquet}
\title{save tidy data.frame as parquet}
\usage{
save_parquet(df_tidy, save_path)
}
\arguments{
\item{df_tidy}{tidy data.frame to save}

\item{save_path}{path to save the file}
}
\description{
save tidy data.frame as parquet
}
