% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyNDB.R
\name{tidy_syohouyaku_prefecture}
\alias{tidy_syohouyaku_prefecture}
\title{処方薬(都道府県別)}
\usage{
tidy_syohouyaku_prefecture(xlsx)
}
\arguments{
\item{xlsx}{excel file to tidy}
}
\description{
都道府県別の処方薬数量データをtidy化
}
