% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyNDB.R
\name{tidy_syohouyaku_sex_age}
\alias{tidy_syohouyaku_sex_age}
\title{Tidying NDB excel file}
\usage{
tidy_syohouyaku_sex_age(xlsx)
}
\arguments{
\item{xlsx}{excel file to tidy}
}
\description{
Tidying NDB excel file
}
