% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyNDB.R
\name{download_xlsx}
\alias{download_xlsx}
\title{ダウンロード}
\usage{
download_xlsx(foldername)
}
\arguments{
\item{foldername}{destination folder path with /}
}
\description{
NDBオープンデータのすべてのxlsxファイルをダウンロードする
}
\note{
参考: https://nigimitama.hatenablog.jp/entry/2018/12/18/173703
}
