% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyNDB.R
\name{tidy_syohouyaku_sex_age}
\alias{tidy_syohouyaku_sex_age}
\title{性年齢別の処方薬数量データをtidy化}
\usage{
tidy_syohouyaku_sex_age(xlsx)
}
\arguments{
\item{xlsx}{excel file to tidy}
}
\description{
性年齢別の処方薬数量データをtidy化
}
